backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
librosa==0.10.1
numpy==1.26.3
soundfile==0.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
git+https://github.com/odysseyml/odyssey-python.git
